+++ docs/changelog.rst
@@ -84,6 +84,13 @@ Documentation
 - Clarified ``skip_diff`` documentation (#1655)
 - Improved documentation relating to validation on import (#1665)
 
+3.3.7 (unreleased)
+------------------
+
+- Pass :meth:`~import_export.mixins.BaseExportMixin.get_export_resource_kwargs` to Resource constructor
+  :meth:`~import_export.admin.ExportMixin.export_action` (#1739)
+- Fix issue with model class passed to Resource constructor crashing on export (#1745)
+
 3.3.6 (2024-01-10)
 ------------------